requirements.txt
streamlit>=1.28.0
crewai>=0.28.8
langchain-openai>=0.0.5
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
plotly>=5.15.0
google-search-results>=2.4.2
tmdbv3api>=1.8.2
google-api-python-client>=2.108.0
python-dotenv>=1.0.0
ragas>=0.0.22
datasets>=2.14.0